pyproject.toml
[project]
name = "Daily Motivational Quote"
version = "0.1.0"
description = "An offline CrewAI project that generates motivational quotes."
authors = [{ name = "Muhammad Thakeeb Muhammad" }]
dependencies = [
    "crewai",
    "python-dotenv"
]

[tool.uv]
dev-dependencies = []
